program operation report updation
