+++ generics/middleware/authenticator.js
@@ -69,7 +69,7 @@ module.exports = function(req, res, next) {
   if (!req.rspObj) req.rspObj = {};
   var rspObj = req.rspObj;
   
-  if((req.path.includes("reports") || req.path.includes("programs")) && req.headers["internal-access-token"] === process.env.INTERNAL_ACCESS_TOKEN) {
+  if((req.path.includes("reports") || req.path.includes("programOperations")) && req.headers["internal-access-token"] === process.env.INTERNAL_ACCESS_TOKEN) {
     req.setTimeout(parseInt(REQUEST_TIMEOUT_FOR_REPORTS));
     next();
     return
